refactor: :art: Refactor done though what asked @WhitedogWarren in his pull review
