+++ src/interfaces/IPrefsUpdate.ts
@@ -1,0 +1,1 @@
+export type PrefsUpdate = { colorScheme: string; }
